requirements.txt
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
